+++ src/InsightMed.Application/Patients/Services/Abstractions/IPatientsService.cs
@@ -5,5 +5,5 @@ namespace InsightMed.Application.Patients.Services.Abstractions;
 public interface IPatientsService
 {
     Task<List<Patient>> GetAllAsync();
-    Task<Patient?> GetByIdAsync(long id);
+    Task<Patient?> GetByIdAsync(int id);
 }
